BWTA now makes use of the static information provided by BWAPI. Added the following 4 functions:
BWTA::readMap - reads in the map data from BWAPI. This allows AI modules to analyze the map in a separate thread.

BaseLocation::getStaticMinerals() - returns all mineral patches associated with the base location, not just accessible ones.

BaseLocation::minerals() - returns the total amount of mineral resources in all accessible mineral patches for the base location. If the mineral patches are not accessible (like an unscouted enemy base), then this will return 0.

BaseLocation::gas() - returns the total amount of gas resources in all accessible vespene geysers for the base location. If the vespene geysers are not accessible (like an unscouted enemy base), then this will return 0.